+++ Pipfile
@@ -15,6 +15,8 @@ sqlalchemy = "*"
 psycopg2-binary = "*"
 python-dotenv = "*"
 boto3 = "*"
+python-multipart = "*"
+pandas = "*"
 
 
 [requires]
