+++ app/main.py
@@ -3,13 +3,11 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 import uvicorn
-from app import models, schemas, routes, database, upload_file
-#from app.database import engine
+from app.database import session_local, engine
 
 # to create our models in the databse
 #models.Base.metadata.create_all(bind=engine)
 from app import models, schemas, routes
-from app.database import session_local, engine
 
 # to create our models in the databse
 models.Base.metadata.create_all(bind=engine)
@@ -25,14 +23,13 @@ To use these interactive docs:
 - Scroll down to see the Server response Code & Details
 """
 
-application = app = FastAPI(
+app = FastAPI(
     title = "Human Rights First Asylum API",
     description = description,
     docs_url='/',
 )
 
 app.include_router(routes.router)
-app.include_router(upload_file.router)
 
 app.add_middleware(
     CORSMiddleware,
