+++ docker-compose.yml
@@ -1,0 +1,9 @@
+version: '3.9'
+services:
+  web:
+    build: ./project
+    command: uvicorn app.main:app --reload --workers 1 --host 0.0.0.0 --port 8000
+    volumes:
+      - ./project:/usr/src/app
+    ports:
+      - 80:8000
